refactor(phase-complete): extract gates, consolidate validation, bound output
